+++ pom.xml
@@ -70,6 +70,11 @@
           <id>simonsators-repo</id>
           <url>https://simonsator.de/repo/</url>
       </repository>
+      <repository>
+          <id>unifiedvision-caesar</id>
+          <name>Unified Vision Repo</name>
+          <url>https://repo.codeblocksmc.de/caesar</url>
+      </repository>
   </repositories>
 
   <dependencies>
@@ -78,11 +83,6 @@
           <artifactId>paper-api</artifactId>
           <version>1.20.4-R0.1-SNAPSHOT</version>
           <scope>provided</scope>
-      </dependency>
-      <dependency>
-          <groupId>io.github.leonardosnt</groupId>
-          <artifactId>bungeechannelapi</artifactId>
-          <version>1.0.0-SNAPSHOT</version>
       </dependency>
       <dependency>
           <groupId>fr.mrmicky</groupId>
@@ -105,6 +105,14 @@
           <groupId>de.simonsator</groupId>
           <artifactId>Party-and-Friends-MySQL-Edition-Spigot-API</artifactId>
           <version>1.6.2-RELEASE</version>
+          <scope>provided</scope>
+      </dependency>
+
+      <!-- Caesar -->
+      <dependency>
+          <groupId>de.julianweinelt.caesar</groupId>
+          <artifactId>CaesarMC</artifactId>
+          <version>1.0.1</version>
           <scope>provided</scope>
       </dependency>
   </dependencies>
