+++ src/main/java/de/codeblocksmc/codelib/api/databsae/DatabaseObject.java
@@ -1,0 +1,4 @@
+package de.codeblocksmc.codelib.api.databsae;
+
+public class DatabaseObject {
+}
