--- src/main/java/de/codeblocksmc/codelib/databsae/DatabaseObject.java
@@ -1,4 +1,0 @@
-package de.codeblocksmc.codelib.databsae;
-
-public class DatabaseObject {
-}
